backend/requirements.txt
sentinelsat==1.2.1
rasterio==1.3.9
numpy==1.24.3
requests==2.31.0
tensorflow==2.15.0
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
matplotlib==3.8.2
scikit-learn==1.4.0
python-dotenv==1.0.0
opencv-python==4.8.1.78